feat : 로또복권묶음 와 당첨번호 를 비교하여 수익률 과당첨된 로또등수 에 따른 갯수를 확인 할 수 있다.

+++ src/test/java/study/lotto/model/IncomeRateTest.java
@@ -2,13 +2,15 @@ package study.lotto.model;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class IncomeRateTest {
 
     @Test
     void 수익과_로또복권을_구매한_비용을_통해서_수익률을_확인할_수_있다() {
-        final IncomeRate incomeRate = IncomeRate.valueOf(5000, 14);
+        final IncomeRate incomeRate = IncomeRate.valueOf(BigDecimal.valueOf(5000), 14);
         assertEquals(incomeRate.getIncomeRate(), 0.35);
     }
 }
